pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "redakt"
version = "0.1.0"
description = "Redakt - Local medical document de-identification powered by Qwen LLM"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "Bora" },
]
keywords = ["anonymization", "deidentification", "medical", "PII", "KVKK", "LLM", "privacy", "redakt"]

dependencies = [
    "PySide6>=6.6",
    "python-docx>=1.1",
    "PyMuPDF>=1.24",
    "openpyxl>=3.1",
    "pydantic>=2.0",
    "httpx>=0.27",
    "Pillow>=10.0",
    "psutil>=5.9",
    "rapidfuzz>=3.0",
]

[project.optional-dependencies]
ollama = [
    "ollama>=0.4",
]
api = [
    "starlette>=0.36",
    "uvicorn[standard]>=0.27",
    "python-multipart>=0.0.9",
]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-qt>=4.3",
    "pyinstaller>=6.0",
    "ruff>=0.5",
    "starlette>=0.36",
    "uvicorn[standard]>=0.27",
    "python-multipart>=0.0.9",
    "ollama>=0.4",
]
ocr = [
    "pytesseract>=0.3",
]

[project.scripts]
redakt = "redakt.__main__:main"

[project.gui-scripts]
redakt-gui = "redakt.__main__:main"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.pytest.ini_options]
asyncio_mode = "auto"
qt_api = "pyside6"
